crop-disease-app/requirements.txt
streamlit
tensorflow
numpy
Pillow










